# [102X] Add CMSSW flags to fastjet(-contrib) compilation, including optimisation & more error checking