@raggleton 2018-11-19
Note that since this adds `-Wall`, it throws up a few warnings generated by HOTVR. They are not serious, and I will fix these in our HOTVR contrib repo.