@raggleton 2018-11-20
BTW if you want to use this, it's probably easiest to setup a fresh 102X area, since it requires recompiling fastjet & CMSSW & UHH2